Integrate soundEffectVolume into Settings system with per-sound bias
- Change soundEffectVolume from Int (0-4) to Float (0.0-1.0) in Settings
- Add SOUND_EFFECT_VOLUME_DEFAULT constant (0.5) to Settings.swift
- Add soundEffectVolume to Codable encode/decode and reset() function
- Update SettingsModalController to manage soundEffectVolume (@Published property)
- Update SettingsView to convert user's 0-4 slider to/from Float value
- Update AudioEffects to read Float directly from Settings.shared
- Add bias property to Sound enum (1.0 default, 0.5 for .click sounds)
- Volume calculation: audioPlayer.volume = sound.bias * soundEffectVolume
- Remove volumeToAVAudioVolume() conversion method (no longer needed)
- soundEffectVolume now part of unified Settings system and included in "Reset All Settings"
- Separates user-facing view concern (0-4 discrete levels) from settings storage (Float 0.0-1.0)

+++ scv-ui/Sources/scvUI/SettingsModalController.swift
@@ -64,7 +64,7 @@ public class SettingsModalController: NSObject, ObservableObject {
     didSet { autosave() }
   }
 
-  @Published var soundEffectVolume: Int {
+  @Published var soundEffectVolume: Float {
     didSet { autosave() }
   }
